use oklog run actors for flag update
